SAMImprovedStreaming v1.1
