+++ Android/SAMImprovedStreaming/ARM/Release/SAMImpro.e187525b.tlog/SAMImprovedStreaming.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=Clang_5_0:VCToolArchitecture=Native32Bit:TargetPlatformVersion=27.0:
+Release|ARM|H:\新建文件夹1\SAMImprovedStreaming\|
