+++ Android/SAMImprovedStreaming/ARM/Release/libSAMImprovedStreaming.Build.CppClean.log
@@ -1,0 +1,21 @@
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\chook.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\cpatch.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\substratedebug.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\substratehook.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\substrateposixmemory.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\canimmanager.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\cstreaming.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\ctimer.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\library.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\inireader.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\minini.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimprovedstreaming.o
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\libsamimprovedstreaming.so
+h:\新建文件夹1\samimprovedstreaming\arm\release\libsamimprovedstreaming.so.recipe
+h:\新建文件夹1\samimprovedstreaming\arm\release\libsamimprovedstreaming.so
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimpro.e187525b.tlog\clang.command.1.tlog
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimpro.e187525b.tlog\microsoft.build.cpptasks.clangcompile.read.1.tlog
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimpro.e187525b.tlog\microsoft.build.cpptasks.clangcompile.write.1.tlog
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimpro.e187525b.tlog\microsoft.build.cpptasks.clanglink.read.1.tlog
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimpro.e187525b.tlog\microsoft.build.cpptasks.clanglink.write.1.tlog
+h:\新建文件夹1\samimprovedstreaming\samimprovedstreaming\arm\release\samimpro.e187525b.tlog\samimprovedstreaming.write.1u.tlog
