+++ PC/LoadWholeMap/IniReader/IniReader.h
