+++ PC/LoadWholeMap/IniReader/LICENSE.txt
